+++ backend/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.2].define(version: 2026_07_26_000001) do
+ActiveRecord::Schema[7.2].define(version: 2026_07_27_000001) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -585,9 +585,12 @@ ActiveRecord::Schema[7.2].define(version: 2026_07_26_000001) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.bigint "unidad_negocio_id"
+    t.bigint "sede_id"
     t.index ["club_id", "clave_sistema"], name: "index_depositos_on_club_id_and_clave_sistema"
+    t.index ["club_id", "sede_id", "clave_sistema"], name: "index_depositos_on_club_sede_clave"
     t.index ["club_id"], name: "index_depositos_on_club_id"
     t.index ["deleted_at"], name: "index_depositos_on_deleted_at"
+    t.index ["sede_id"], name: "index_depositos_on_sede_id"
     t.index ["unidad_negocio_id"], name: "index_depositos_on_unidad_negocio_id"
   end
 
@@ -2178,6 +2181,7 @@ ActiveRecord::Schema[7.2].define(version: 2026_07_26_000001) do
   add_foreign_key "cuenta_corrientes", "pacientes"
   add_foreign_key "cuenta_corrientes", "users", column: "deleted_by_id"
   add_foreign_key "depositos", "clubs"
+  add_foreign_key "depositos", "sedes"
   add_foreign_key "depositos", "unidades_negocio", column: "unidad_negocio_id"
   add_foreign_key "dispensacion_items", "dispensaciones", column: "dispensacion_id"
   add_foreign_key "dispensacion_items", "stocks"